% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepLast.R
\name{keepLast}
\alias{keepLast}
\title{Plot a Raster* object with terrain and bathymetry}
\usage{
keepLast(...)
}
\arguments{
\item{...}{Any argument that can be passed to \code{\link[base]{dir}}}

\item{path}{a character vector of full path names; the default corresponds to the working directory,
\code{\link[base]{getwd}}. Tilde expansion (see \code{\link[base]{path.expand}}) is performed. Missing values will be ignored.}
}
\description{
This function is a wrapper of  \code{\link[raster]{spplot}}
}
\examples{
keepLast("D:/Projects_MOST/GPSC_cruise/OR1_1219/scoc1219/measurement")
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
