% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touPlot.R
\name{touPlot}
\alias{touPlot}
\title{Plot TOU}
\usage{
touPlot(x)
}
\arguments{
\item{x}{a dataframe derived from \code{\link[PreSens]{keepLast}}}
}
\description{
This function plots TOU and the 85\% of the initial DO value
}
\examples{
dat <- keepLast("D:/Projects_MOST/GPSC_cruise/OR1_1219/scoc1219/measurement")
touPlot(dat)
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
