% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepLast.R
\name{keepLast}
\alias{keepLast}
\title{Extract the last observation of each time series}
\usage{
keepLast()
}
\description{
This function the last observation of each time series from PreSens logging data
}
\details{
Set the working directory to the directory of PreSens data using \code{\link[base]{setwd}}
}
\examples{
keepLast()
}
\author{
Chih-Lin Wei <chihlinwei@gmail.com>
}
